Move web worker and types into @pmndrs/cannon-worker-api package
