+++ packages/cannon-worker-api/src/index.ts
@@ -1,1 +1,4 @@
-export const placeholderExport = 'placeholder'
+export * from './body'
+export * from './cannon-worker-api'
+export * from './props-to-body'
+export * from './setup'
