+++ packages/cannon-worker-api/src/worker.d.ts
@@ -1,0 +1,4 @@
+declare module 'web-worker:*' {
+  const WorkerFactory: new () => Worker
+  export default WorkerFactory
+}
