+++ packages/cannon-worker-api/src/worker/operations/add-raycast-vehicle.ts
