+++ packages/cannon-worker-api/src/worker/state.ts
