+++ packages/react-three-cannon/src/hooks.ts
@@ -1,108 +1,47 @@
-import type { ContactMaterialOptions, MaterialOptions } from 'cannon-es'
-import type { DependencyList, MutableRefObject, Ref, RefObject } from 'react'
-import { useContext, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react'
-import { DynamicDrawUsage, Euler, InstancedMesh, MathUtils, Object3D, Quaternion, Vector3 } from 'three'
-
-import type { CannonWorkerAPI } from './cannon-worker-api'
 import type {
   AtomicName,
-  CollideBeginEvent,
-  CollideEndEvent,
-  CollideEvent,
+  AtomicProps,
+  BodyProps,
+  BodyShapeType,
+  BoxProps,
+  CannonWorkerAPI,
+  CompoundBodyProps,
   ConeTwistConstraintOpts,
   ConstraintOptns,
   ConstraintTypes,
+  ConvexPolyhedronArgs,
+  ConvexPolyhedronProps,
+  CylinderProps,
   DistanceConstraintOpts,
+  HeightfieldProps,
   HingeConstraintOpts,
   LockConstraintOpts,
+  ParticleProps,
+  PlaneProps,
   PointToPointConstraintOpts,
   PropValue,
-  ProviderContext,
   Quad,
   RayhitEvent,
   RayMode,
   RayOptions,
   SetOpName,
+  SphereArgs,
+  SphereProps,
   SpringOptns,
   SubscriptionName,
   SubscriptionTarget,
+  TrimeshProps,
   Triplet,
   VectorName,
   WheelInfoOptions,
-} from './setup'
+} from '@pmndrs/cannon-worker-api'
+import type { ContactMaterialOptions, MaterialOptions } from 'cannon-es'
+import type { DependencyList, MutableRefObject, Ref, RefObject } from 'react'
+import { useContext, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react'
+import { DynamicDrawUsage, Euler, InstancedMesh, MathUtils, Object3D, Quaternion, Vector3 } from 'three'
+
+import type { ProviderContext } from './setup'
 import { context, debugContext } from './setup'
-
-export type AtomicProps = {
-  allowSleep: boolean
-  angularDamping: number
-  collisionFilterGroup: number
-  collisionFilterMask: number
-  collisionResponse: boolean
-  fixedRotation: boolean
-  isTrigger: boolean
-  linearDamping: number
-  mass: number
-  material: MaterialOptions
-  sleepSpeedLimit: number
-  sleepTimeLimit: number
-  userData: Record<PropertyKey, any>
-}
-
-export type VectorProps = Record<VectorName, Triplet>
-type VectorTypes = Vector3 | Triplet
-
-export type BodyProps<T extends any[] = unknown[]> = Partial<AtomicProps> &
-  Partial<VectorProps> & {
-    args?: T
-    onCollide?: (e: CollideEvent) => void
-    onCollideBegin?: (e: CollideBeginEvent) => void
-    onCollideEnd?: (e: CollideEndEvent) => void
-    quaternion?: Quad
-    rotation?: Triplet
-    type?: 'Dynamic' | 'Static' | 'Kinematic'
-  }
-
-export type BodyPropsArgsRequired<T extends any[] = unknown[]> = BodyProps<T> & {
-  args: T
-}
-
-export type ShapeType =
-  | 'Box'
-  | 'ConvexPolyhedron'
-  | 'Cylinder'
-  | 'Heightfield'
-  | 'Particle'
-  | 'Plane'
-  | 'Sphere'
-  | 'Trimesh'
-export type BodyShapeType = ShapeType | 'Compound'
-
-export type CylinderArgs = [radiusTop?: number, radiusBottom?: number, height?: number, numSegments?: number]
-export type SphereArgs = [radius: number]
-export type TrimeshArgs = [vertices: ArrayLike<number>, indices: ArrayLike<number>]
-export type HeightfieldArgs = [
-  data: number[][],
-  options: { elementSize?: number; maxValue?: number; minValue?: number },
-]
-export type ConvexPolyhedronArgs<V extends VectorTypes = VectorTypes> = [
-  vertices?: V[],
-  faces?: number[][],
-  normals?: V[],
-  axes?: V[],
-  boundingSphereRadius?: number,
-]
-
-export type PlaneProps = BodyProps
-export type BoxProps = BodyProps<Triplet>
-export type CylinderProps = BodyProps<CylinderArgs>
-export type ParticleProps = BodyProps
-export type SphereProps = BodyProps<SphereArgs>
-export type TrimeshProps = BodyPropsArgsRequired<TrimeshArgs>
-export type HeightfieldProps = BodyPropsArgsRequired<HeightfieldArgs>
-export type ConvexPolyhedronProps = BodyProps<ConvexPolyhedronArgs>
-export interface CompoundBodyProps extends BodyProps {
-  shapes: BodyProps & { type: ShapeType }[]
-}
 
 export type AtomicApi<K extends AtomicName> = {
   set: (value: AtomicProps[K]) => void
