+++ packages/react-three-cannon/src/useUpdateWorldPropsEffect.ts
@@ -1,8 +1,7 @@
+import type { CannonWorkerAPI, WorldPropName } from '@pmndrs/cannon-worker-api'
 import { useEffect } from 'react'
 
-import type { CannonWorkerAPI } from './cannon-worker-api'
 import type { ProviderProps } from './Provider'
-import type { WorldPropName } from './setup'
 
 type Props = Pick<Required<ProviderProps>, WorldPropName> & { worker: CannonWorkerAPI }
